requirements.txt
altgraph==0.17
asgiref==3.3.1
auto-py-to-exe==2.7.8
bottle==0.12.18
bottle-websocket==0.2.9
dj-database-url==0.5.0
Django==3.0
django-classy-tags==2.0.0
django-cms==3.7.4
django-filer==2.0.2
django-filter==2.4.0
django-formtools==2.2
django-jet==1.0.8
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==3.0.0
django-sekizai==2.0.0
django-treebeard==4.3.1
djangocms-admin-style==2.0.0
djangocms-attributes-field==2.0.0
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-icon==2.0.0
djangocms-installer==1.2.3
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
djangorestframework==3.12.2
djangorestframework-datatables==0.5.2
easy-thumbnails==2.7
Eel==0.12.4
feedparser==6.0.2
future==0.18.2
gevent==20.9.0
gevent-websocket==0.10.1
google-api-python-client==1.4.1
greenlet==0.4.17
html5lib==1.1
httplib2==0.18.1
importlib-metadata==2.0.0
Markdown==3.3.3
mysqlclient==2.0.1
oauth2client==4.1.3
Pillow==8.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.2
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
pytz==2020.1
rsa==4.6
sgmllib3k==1.0.0
six==1.15.0
South==1.0.2
sqlparse==0.4.1
tzlocal==2.1
Unidecode==1.1.1
uritemplate==3.0.1
webencodings==0.5.1
whichcraft==0.6.1
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.1.2
